Add sdk rule for supported versions (#5)
We do get issues on the SDKs where people were using really old versions
of things that we no longer support, so I think is valid to also have
rules for the implementation itself.
This PR updates the template to reflect a few examples of components and
also create a new rule for dependency on SDKs.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

- **New Features**
- Introduced a new rule to detect if language and runtime versions in
SDK dependencies are supported, helping ensure compatibility and
up-to-date usage.
- **Documentation**
- Updated rule template to allow rules to target SDKs, Collectors, and
other components in addition to OTLP elements.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->